fix entity dimensions
