+++ common/src/main/resources/walkers.mixins.json
@@ -6,6 +6,7 @@
   "mixins": [
     "AbstractHorseMixin",
     "ActiveTargetGoalMixin",
+    "AvatarMixin",
     "AvoidEntityGoalMixin",
     "AxolotlAttackablesSensorMixin",
     "ConsumableMixin",
